13:20 --- 28/01/2023

+++ IntellaQeust.BusinessLogic/Categories/CategoryService.cs
@@ -1,34 +1,33 @@
-﻿using IntellaQeust.BusinessLogic.CategoryModels;
+﻿using IntellaQeust.BusinessLogic.CategoryModel;
 using IntellaQeust.BusinessLogic.Exceptions;
 using IntellaQeust.BusinessLogic.Exceptions.ExceptionMassages;
 using IntellaQuest.BusinessLogic.Mappers;
-using IntellaQuest.Domain;
-using IntellaQuest.Repository;
-using IntellaQuest.Repository.Repositories;
+using IntellaQuest.Data.NHibernate.ConfigurationRepository;
+using IntellaQuest.Data.NHibernate.Repositories;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 
-namespace IntellaQeust.BusinessLogic.Services
+namespace IntellaQeust.Business.Services
 {
-    public interface ICategoriesService
+    public interface ICategoryService
     {
-        CategoriesViewModel Get(Guid Id);
-        List<CategoriesViewModel> GetAll();
-        Guid Create(CategoriesViewModel model);
-        void Update(CategoriesViewModel model);
+        CategoryViewModel Get(Guid Id);
+        List<CategoryViewModel> GetAll();
+        Guid Create(CategoryViewModel model);
+        void Update(CategoryViewModel model);
         void DeleteById(Guid Id);
-        void Delete(CategoriesViewModel model);
         bool CheckCategoryNameExists(String Name);
         bool CheckCategoryStatus(bool status);
     }
 
-    public class CategoriesService : ICategoriesService
+    public class CategoryService : ICategoryService
     {
-        private readonly ICategoriesRepository _categoryRepository;
+        private readonly ICategoryRepository _categoryRepository;
         private readonly IUnitOfWork _unitOfWork;
 
-        public CategoriesService(ICategoriesRepository categoryRepository, IUnitOfWork unitOfWork)
+        public CategoryService(ICategoryRepository categoryRepository, IUnitOfWork unitOfWork)
         {
             _categoryRepository = categoryRepository;
             _unitOfWork = unitOfWork;
@@ -37,54 +36,49 @@ namespace IntellaQeust.BusinessLogic.Services
 
         public bool CheckCategoryNameExists(String Name)
         {
-            return _categoryRepository.FilterBy(x=>x.Name==Name).ToList().Any();
+            return _categoryRepository.FilterBy(x=>x.Name==Name).Any();
         }
         public bool CheckCategoryStatus(bool status)
         {
             return _categoryRepository.FilterBy(x => x.Status == status).ToList().Any();
         }
 
-        public Guid Create(CategoriesViewModel model)
+        public Guid Create(CategoryViewModel model)
         {
             using (_unitOfWork.BeginTransaction())
             {
                 if (CheckCategoryNameExists(model.Name))
                 {
-                    throw new BllException(ShopExceptionsMassages.CategoriesExceptionMassages.NAME_ALREADY_EXIST_EXCEPTION);
+                    throw new BllException(ShopExceptionMassages.CategoriesExceptionMassages.NAME_ALREADY_EXIST_EXCEPTION);
                 }
-                var categoryEntity = model.MapToModel();
                 _categoryRepository.Add(model.MapToModel());
                 _unitOfWork.Commit();
-                return categoryEntity.Id;
+
+                return model.Id;
             }
         }
-
-        public void Delete(CategoriesViewModel model)
-        {
-            using (_unitOfWork.BeginTransaction())
-            {
-                _categoryRepository.Delete(model.MapToModel());
-                _unitOfWork.Commit();
-            }
-        }
-
         public void DeleteById(Guid Id)
         {
             using (_unitOfWork.BeginTransaction())
             {
+                var categoryEntity = Get(Id);
+                if (_categoryRepository.FindBy(Id) == null)
+                {
+                    throw new BllException(ShopExceptionMassages.CategoriesExceptionMassages.NOT_FOUND_EXCEPTION);
+                }
                 _categoryRepository.Delete(Id);
                 _unitOfWork.Commit();
             }
         }
 
-        public CategoriesViewModel Get(Guid Id)
+        public CategoryViewModel Get(Guid Id)
         {
             using (_unitOfWork.BeginTransaction())
             {
                 var category = _categoryRepository.FindBy(Id);
                 if(category == null)
                 {
-                    throw new BllException(ShopExceptionsMassages.CategoriesExceptionMassages.NOT_FOUND_EXCEPTION);
+                    throw new BllException(ShopExceptionMassages.CategoriesExceptionMassages.NOT_FOUND_EXCEPTION);
                 }
                 _unitOfWork.Commit();
                 _unitOfWork.Rollback();
@@ -93,7 +87,7 @@ namespace IntellaQeust.BusinessLogic.Services
             }
         }
 
-        public List<CategoriesViewModel> GetAll()
+        public List<CategoryViewModel> GetAll()
         {
             using (_unitOfWork.BeginTransaction())
             {
@@ -103,23 +97,21 @@ namespace IntellaQeust.BusinessLogic.Services
             }
         }
 
-        public void Update(CategoriesViewModel model)
+        public void Update(CategoryViewModel model)
         {
             using(_unitOfWork.BeginTransaction())
             {
                 var category = _categoryRepository.FindBy(model.Id);
                 if (category == null)
                 {
-                    throw new BllException(ShopExceptionsMassages.CustomerExceptionMassages.NOT_FOUND_EXCEPTION);
+                    throw new BllException(ShopExceptionMassages.CustomerExceptionMassages.NOT_FOUND_EXCEPTION);
                 }
-                // TODO : if not found throw exception
 
                 category.Name= model.Name;
                 category.Status = model.Status;
 
                 _categoryRepository.Update(category);
                 _unitOfWork.Commit();
-
             }
         }
     }
